scripts/requirements.txt
astroid==2.4.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
dateparser==1.0.0
DateTime==4.3
google-api-core==1.17.0
google-auth==1.16.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
pip==20.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.28
requests==2.23.0
rsa==4.0
setuptools==46.4.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
wheel==0.34.2
wrapt==1.12.1
zope.interface==5.2.0